backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
celery[redis]==5.4.0
redis==5.2.1
pydantic-settings==2.7.1
httpx==0.28.1
requests==2.32.3
pytest==8.3.4
pytest-asyncio==0.25.0
websockets==14.1
PyJWT[crypto]==2.9.0
gunicorn==23.0.0
sentry-sdk[fastapi]==2.19.2
stripe==11.4.1
resend==2.5.0
jinja2==3.1.5
yfinance==0.2.50
anthropic==0.43.0
weasyprint==62.3
openpyxl==3.1.5
